WebsiteQLKS_Django_React/hotelAPI/requirements.txt
appnope==0.1.4
asgiref==3.8.1
asttokens==2.4.1
backcall==0.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.40.0
colorama==0.4.6
cryptography==42.0.8
decorator==5.1.1
Django==5.0.6
django-oauth-toolkit==2.4.0
djangorestframework==3.15.2
drf-yasg==1.21.7
executing==2.0.1
idna==3.7
inflection==0.5.1
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.4
jwcrypto==1.5.6
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
oauthlib==3.2.2
packaging==24.1
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.1.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.18.0
PyMySQL==1.1.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.2.0
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
tornado==6.3.3
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
wcwidth==0.2.13
